reviews_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
motor==3.7.1
mypy_extensions==1.1.0
orjson==3.10.18
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
